+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Move Redis DB persister to extras package
     - Remove Loda estimate bins static method
     - Change Grid Search base estimator param name to class
+    - Remove Dataset cast to string preview
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features